[Feat] TrainerUseCase 작성
